requirements.txt
numpy==1.24.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.21.4
transformers==4.33.3
accelerate==0.24.1
safetensors==0.4.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
huggingface-hub==0.17.3
scipy==1.11.4